refactor: extract backup runner
